Austin's AI Agent Teams/programmatic_blog_generator/requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.2
beautifulsoup4>=4.12.2
markdown>=3.5.1
python-slugify>=8.0.1
pytest>=7.4.3
black>=23.10.1
isort>=5.12.0
flake8>=6.1.0 